+++ src/features/calendar/shared/api/calendarApi.ts
@@ -1,33 +1,20 @@
-import calendarScheduleMock from '../../mocks/CalendarScheduleMock';
-import { CalendarSchedule, CalendarScheduleRequest } from '../../types';
+import { CalendarScheduleRequest, CalendarScheduleResponse } from '../../types';
 
-import dayjs from '@/shared/lib/dayjs';
+import { axiosInstance } from '@/shared/api/axiosInstance';
 
-const fetchCalendarSchedule = async ({
+export const calendarScheduleApi = async ({
   startDate,
   endDate,
-}: CalendarScheduleRequest): Promise<CalendarSchedule[]> => {
-  // AI 사용
-  // mock 데이터로 대체
-  const mock = calendarScheduleMock;
+}: CalendarScheduleRequest): Promise<CalendarScheduleResponse> => {
+  const response = await axiosInstance.get<CalendarScheduleResponse>(
+    '/schedules',
+    {
+      params: {
+        startDate,
+        endDate,
+      },
+    },
+  );
 
-  // 1초 딜레이 시뮬레이션
-  await new Promise(resolve => setTimeout(resolve, 1000));
-
-  // startDate와 endDate 사이의 스케줄 필터링
-  const filteredSchedules = mock.data.filter(schedule => {
-    const scheduleStart = dayjs(schedule.startDate);
-    const scheduleEnd = dayjs(schedule.endDate);
-    const requestStart = dayjs(startDate);
-    const requestEnd = dayjs(endDate);
-
-    // 스케줄 기간이 요청 기간과 겹치는지 확인
-    return (
-      scheduleStart.isBefore(requestEnd) && scheduleEnd.isAfter(requestStart)
-    );
-  });
-
-  return filteredSchedules;
+  return response.data;
 };
-
-export default fetchCalendarSchedule;
